backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
prometheus-client==0.20.0
psutil==5.9.8
anthropic==0.25.0
openai==1.25.0
httpx==0.27.0
pydantic==2.7.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
numpy==1.26.4
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
